move contents of container_umx.py to mi_tracker_umx.py
